Add NewElement component
