+++ src/components/FormElementsEditor/components/NewElement/index.ts
@@ -1,0 +1,1 @@
+export * from './NewElement';
