+++ src/components/FormElementsEditor/components/index.ts
@@ -1,1 +1,2 @@
 export * from './ElementEditor';
+export * from './NewElement';
